Revert "Main rebased (#11)"
This reverts commit c84662b31a8fa4d24cce840de71d5523911e5c81.

+++ airbyte-integrations/connectors/source-amazon-ads/source_amazon_ads/schemas/sponsored_brands.py
@@ -51,7 +51,7 @@ class BrandsAdGroup(CatalogModel):
     campaignId: Decimal
     adGroupId: Decimal
     name: str
-    bid: Decimal
+    bid: int
     keywordId: Decimal
     keywordText: str
     nativeLanguageKeyword: str
